move metrics to hapi service
