+++ packages/relay/src/lib/clients/sdkClient.ts
@@ -79,12 +79,6 @@ export class SDKClient {
     private readonly logger: Logger;
 
     /**
-     * The metrics register used for metrics tracking.
-     * @private
-     */
-    private readonly register: Registry;
-
-    /**
      * This limiter tracks hbar expenses and limits.
      * @private
      */
@@ -99,10 +93,9 @@ export class SDKClient {
     private consensusNodeClientHistogramCost;
     private consensusNodeClientHistogramGasFee;
     private operatorAccountGauge;
-    private operatorAccountId;
 
     // populate with consensusnode requests via SDK
-    constructor(clientMain: Client, logger: Logger, register: Registry, hbarLimiter: HbarLimit) {
+    constructor(clientMain: Client, logger: Logger, hbarLimiter: HbarLimit, metrics: any) {
         this.clientMain = clientMain;
 
         if (process.env.CONSENSUS_MAX_EXECUTION_TIME) {
@@ -112,48 +105,9 @@ export class SDKClient {
         }
 
         this.logger = logger;
-        this.register = register;
-        this.operatorAccountId = clientMain.operatorAccountId ? clientMain.operatorAccountId.toString() : 'UNKNOWN';
-
-        // clear and create metrics in registry
-        const metricHistogramCost = 'rpc_relay_consensusnode_response';
-        register.removeSingleMetric(metricHistogramCost);
-        this.consensusNodeClientHistogramCost = new Histogram({
-            name: metricHistogramCost,
-            help: 'Relay consensusnode mode type status cost histogram',
-            labelNames: ['mode', 'type', 'status', 'caller', 'interactingEntity'],
-            registers: [register]
-        });
-        const metricHistogramGasFee = 'rpc_relay_consensusnode_gasfee';
-        register.removeSingleMetric(metricHistogramGasFee);
-        this.consensusNodeClientHistogramGasFee = new Histogram({
-            name: metricHistogramGasFee,
-            help: 'Relay consensusnode mode type status gas fee histogram',
-            labelNames: ['mode', 'type', 'status', 'caller', 'interactingEntity'],
-            registers: [register]
-        });
-
-        const metricGaugeName = 'rpc_relay_operator_balance';
-        register.removeSingleMetric(metricGaugeName);
-        this.operatorAccountGauge = new Gauge({
-            name: metricGaugeName,
-            help: 'Relay operator balance gauge',
-            labelNames: ['mode', 'type', 'accountId'],
-            registers: [register],
-            async collect() {
-                // Invoked when the registry collects its metrics' values.
-                // Allows for updated account balance tracking
-                try {
-                    const accountBalance = await (new AccountBalanceQuery()
-                        .setAccountId(clientMain.operatorAccountId!))
-                        .execute(clientMain);
-                    this.labels({ 'accountId': clientMain.operatorAccountId!.toString() })
-                        .set(accountBalance.hbars.toTinybars().toNumber());
-                } catch (e: any) {
-                    logger.error(e, `Error collecting operator balance. Skipping balance set`);
-                }
-            },
-        });
+        this.consensusNodeClientHistogramCost = metrics.costHistogram;
+        this.consensusNodeClientHistogramGasFee = metrics.gasHistogram;
+        this.operatorAccountGauge = metrics.operatorGauge;
 
         this.hbarLimiter = hbarLimiter;
         this.cache = new LRU({ max: constants.CACHE_MAX, ttl: constants.CACHE_TTL.ONE_HOUR });
